refactor: mock data
