+++ frontend/rust-lib/flowy-database/src/services/group/mod.rs
@@ -5,7 +5,6 @@ mod controller_impls;
 mod entities;
 mod group_util;
 
-pub(crate) use action::*;
 pub(crate) use configuration::*;
 pub(crate) use controller::*;
 pub(crate) use controller_impls::*;
