+++ frontend/rust-lib/flowy-database/tests/grid/block_test/script.rs
@@ -1,6 +1,6 @@
 use crate::grid::block_test::script::RowScript::{AssertCell, CreateRow};
 use crate::grid::block_test::util::GridRowTestBuilder;
-use crate::grid::grid_editor::GridEditorTest;
+use crate::grid::database_editor::DatabaseEditorTest;
 use flowy_database::entities::{CellPathParams, CreateRowParams, DatabaseViewLayout, FieldType, RowPB};
 use flowy_database::services::field::*;
 use flowy_database::services::row::DatabaseBlockRow;
@@ -48,13 +48,13 @@ pub enum RowScript {
     },
 }
 
-pub struct GridRowTest {
-    inner: GridEditorTest,
+pub struct DatabaseRowTest {
+    inner: DatabaseEditorTest,
 }
 
-impl GridRowTest {
+impl DatabaseRowTest {
     pub async fn new() -> Self {
-        let editor_test = GridEditorTest::new_table().await;
+        let editor_test = DatabaseEditorTest::new_table().await;
         Self { inner: editor_test }
     }
 
@@ -282,15 +282,15 @@ fn block_from_row_pbs(row_orders: Vec<RowPB>) -> Vec<DatabaseBlockRow> {
     map.into_values().collect::<Vec<_>>()
 }
 
-impl std::ops::Deref for GridRowTest {
-    type Target = GridEditorTest;
+impl std::ops::Deref for DatabaseRowTest {
+    type Target = DatabaseEditorTest;
 
     fn deref(&self) -> &Self::Target {
         &self.inner
     }
 }
 
-impl std::ops::DerefMut for GridRowTest {
+impl std::ops::DerefMut for DatabaseRowTest {
     fn deref_mut(&mut self) -> &mut Self::Target {
         &mut self.inner
     }
@@ -303,7 +303,7 @@ pub struct CreateRowScriptBuilder<'a> {
 }
 
 impl<'a> CreateRowScriptBuilder<'a> {
-    pub fn new(test: &'a GridRowTest) -> Self {
+    pub fn new(test: &'a DatabaseRowTest) -> Self {
         Self {
             builder: test.row_builder(),
             data_by_field_type: HashMap::new(),
