+++ frontend/rust-lib/flowy-database/tests/grid/mock_data/mod.rs
@@ -1,0 +1,19 @@
+mod board_mock_data;
+mod calendar_mock_data;
+mod grid_mock_data;
+
+pub use board_mock_data::*;
+pub use calendar_mock_data::*;
+pub use grid_mock_data::*;
+
+pub const GOOGLE: &str = "Google";
+pub const FACEBOOK: &str = "Facebook";
+pub const TWITTER: &str = "Twitter";
+
+pub const COMPLETED: &str = "Completed";
+pub const PLANNED: &str = "Planned";
+pub const PAUSED: &str = "Paused";
+
+pub const FIRST_THING: &str = "Wake up at 6:00 am";
+pub const SECOND_THING: &str = "Get some coffee";
+pub const THIRD_THING: &str = "Start working";
