+++ frontend/rust-lib/flowy-database/tests/grid/mod.rs
@@ -1,8 +1,10 @@
 mod block_test;
 mod cell_test;
+mod database_editor;
 mod field_test;
 mod filter_test;
-mod grid_editor;
 mod group_test;
 mod snapshot_test;
 mod sort_test;
+
+mod mock_data;
